8-9: Ensure the AttributeMenu can be closed
We want AttributeMenu to be independent from WBP_Overlay.
So we create an event dispatcher in AttributeMenu called AttributeMenuClosed.
In WBP_Overlay, when we create AttributeMenu, we assign an event to the dispatcher to listen for it.
When triggered, it'll make the attribute menu button enabled again
(since it's disabled the moment it's clicked to open the attribute menu).

IMPORTANT:
Since we want the AttributeMenu to center on the screen, this differs from the course
in that the AttributeMenu's Overlay_Root is wrapped in a Canvas Panel and we set the
anchors to center and adjusted the alignment to 0.5x0.5.
This way, we don't have to set the position in viewport when we create the attribute menu
but simply create it and add it to the viewport so that it is automatically centered.