8-17: Construct the AttributeMenuWidgetController
We construct the attribute menu widget controller (based on its class)
in AuraHUD and implement a getter function.
Then, inside our blueprint function library, we implement a getter for
the attribute menu widget controller that uses AuraHUD's function.
After that, in UE editor, in the event graph of  WBP_AttributeMenu,
we use the blueprint function library getter to set the widget controller
for the WBP_AttributeMenu.